+++ camomilla/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "5.4.2"
+__version__ = "5.7.1"
 
 
 def get_core_apps():